HW01 corrections after review finally

+++ src/main/java/ru/javawebinar/topjava/web/MealServlet.java
@@ -32,18 +32,19 @@ public class MealServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String forward = null;
         String action = request.getParameter("action");
-        if
-        if (action.equalsIgnoreCase("delete")) {
+        if (action != null && !action.isEmpty() && action.equalsIgnoreCase("delete")) {
             int mealToId = Integer.parseInt(request.getParameter("mealToId"));
             mealToDao.deleteMealTo(mealToId);
             log.debug("forward to meals from delete");
-        } else if (action.equalsIgnoreCase("edit")) {
+            response.sendRedirect("meals");
+            return;
+        } else if (action != null && !action.isEmpty() && action.equalsIgnoreCase("edit")) {
             forward = "/meal.jsp";
             int mealToId = Integer.parseInt(request.getParameter("mealToId"));
             MealTo mealTo = mealToDao.getMealToById(mealToId);
             request.setAttribute("mealTo", mealTo);
             log.debug("forward to meal from edit");
-        } else if (action.equalsIgnoreCase("addMeal")) {
+        } else if (action != null && !action.isEmpty() && action.equalsIgnoreCase("addMeal")) {
             forward = "/meal.jsp";
             Meal meal = new Meal(LocalDateTime.now().withNano(0).withSecond(0), null, 0);
             request.setAttribute("mealTo", meal);
